+++ .github/workflows/ci.yml
@@ -2,52 +2,63 @@ name: CI/CD Pipeline
 
 on:
   push:
-    branches: [master]
+    branches: [master, main]
   pull_request:
-    branches: [main, master]
+    branches: [master, main]
 
 jobs:
   test:
-    name: Test and Build
+    name: Lint, type-check, spell, build
     runs-on: ubuntu-latest
 
     steps:
       - name: Checkout repository
         uses: actions/checkout@v4
 
+      - name: Setup pnpm
+        uses: pnpm/action-setup@v4
+        with:
+          version: 9
+
       - name: Setup Node.js
-        uses: oven-sh/setup-bun@v2
+        uses: actions/setup-node@v4
+        with:
+          node-version: 20
+          cache: pnpm
 
       - name: Install dependencies
-        run: bun install
+        run: pnpm install --frozen-lockfile
 
-      - name: Run lint, type-check, and build
-        run: bun run test
+      - name: Format check
+        run: pnpm run format:check
 
-      - name: Test build output
+      - name: Lint
+        run: pnpm run lint
+
+      - name: Type check
+        run: pnpm run type-check
+
+      - name: Spell check
+        run: pnpm run spell
+
+      - name: Build
+        run: pnpm run build
+
+      - name: Verify build output
         run: |
           if [ ! -d ".next" ]; then
-            echo "❌ Build failed - .next directory not found"
+            echo "Build failed - .next directory not found"
             exit 1
           fi
-          echo "✅ Build successful!"
-
-      - name: Upload build artifacts
-        uses: actions/upload-artifact@v4
-        with:
-          name: build-files
-          path: .next/
-          retention-days: 1
+          echo "Build successful"
 
   deploy-check:
     name: Deployment Check
     runs-on: ubuntu-latest
     needs: test
-    if: github.ref == 'refs/heads/master'
+    if: github.ref == 'refs/heads/master' || github.ref == 'refs/heads/main'
 
     steps:
       - name: Deployment Status
         run: |
-          echo "🚀 Ready for deployment!"
-          echo "Vercel will automatically deploy this build when pushed to master"
-          echo "Visit your Vercel dashboard to monitor deployment status"
+          echo "Ready for deployment - Vercel will auto-deploy this build"
