+++ src/lib/searchConfig.ts
@@ -1,5 +1,4 @@
-import MiniSearch from "minisearch";
-import { searchData } from "@/data/searchData";
+import type MiniSearch from "minisearch";
 
 export interface SearchResult {
   id: string;
@@ -15,14 +14,13 @@ export interface SearchResult {
   };
 }
 
-// Generic interface for MiniSearch results (contains id, score, and all stored fields)
 interface MiniSearchRawResult {
   id: string;
   score: number;
   match?: {
     [key: string]: string[];
   };
-  [key: string]: unknown; // Allow any additional stored fields
+  [key: string]: unknown;
 }
 
 export const mapMiniSearchResults = (
@@ -41,18 +39,44 @@ export const mapMiniSearchResults = (
   }));
 };
 
-export const createSearchInstance = () => {
-  const miniSearch = new MiniSearch({
-    fields: ["title", "description", "content", "tags", "category"],
-    storeFields: ["title", "description", "content", "url", "category", "tags"],
-    searchOptions: {
-      boost: { title: 2, tags: 1.5, description: 1.2 },
-      fuzzy: 0.2,
-      prefix: true,
-      combineWith: "OR",
-    },
-  });
+let cachedInstance: MiniSearch | null = null;
+let pendingInstance: Promise<MiniSearch> | null = null;
 
-  miniSearch.addAll(searchData);
-  return miniSearch;
+// Lazily load both MiniSearch and the search-data JSON only on first call.
+// Subsequent callers receive the same instance — the index is built once per
+// session, not on every SearchBar mount.
+export const getSearchInstance = (): Promise<MiniSearch> => {
+  if (cachedInstance) return Promise.resolve(cachedInstance);
+  if (pendingInstance) return pendingInstance;
+
+  pendingInstance = (async () => {
+    const [{ default: MiniSearchCtor }, { searchData }] = await Promise.all([
+      import("minisearch"),
+      import("@/data/searchData"),
+    ]);
+
+    const miniSearch = new MiniSearchCtor({
+      fields: ["title", "description", "content", "tags", "category"],
+      storeFields: [
+        "title",
+        "description",
+        "content",
+        "url",
+        "category",
+        "tags",
+      ],
+      searchOptions: {
+        boost: { title: 2, tags: 1.5, description: 1.2 },
+        fuzzy: 0.2,
+        prefix: true,
+        combineWith: "OR",
+      },
+    });
+
+    miniSearch.addAll(searchData);
+    cachedInstance = miniSearch;
+    return miniSearch;
+  })();
+
+  return pendingInstance;
 };
